Deacidify preferred versions
