+++ src/Distribution/Server/Features/ReverseDependencies/State.hs
@@ -47,6 +47,7 @@ import           Distribution.Package
 import           Distribution.PackageDescription
 import           Distribution.Server.Packages.Types
 import           Distribution.Server.Framework.MemSize
+import           Distribution.Server.Features.PreferredVersions (maybeBestVersion)
 import           Distribution.Server.Features.PreferredVersions.State
 import           Distribution.Server.Packages.PackageIndex (PackageIndex)
 import qualified Distribution.Server.Packages.PackageIndex as PackageIndex
